Revert "add multi tenant testcases"
This reverts commit f7791b64cff695c5f40b6ef49fb8b2f6878f5311.
